// apps/stm-g/src/modules-features/admin/ModuleCertificate/CertificateDecision/Interfaces/ICertificateResult.ts
export interface ICertificateResult {
    userId?: number;
    examId?: number;
    decisionNumber?: string;
    decisionDate?: string;
    certificateNumber?: string;
    receivedDate?: string;
    registrationNumber?: string;
    handoverStatus?: number;
    note?: string;
    isPass?: boolean;
    point?: number | null;
    certificateDecisionId?: number | null;
    user?: {
        id?: number;
        isBlocked?: boolean;
        roleId?: number;
        code?: string;
        email?: string;
        phoneNumber?: string;
        address?: string;
        avatarPath?: string;
        fullName?: string;
        gender?: number;
        dateOfBirth?: string;
        educationLevel?: number;
        courseSectionStudentPoints?: any[] | null;
    };
    exam?: {
        programId?: number;
        examDate?: string;
        roomTypeId?: number;
        status?: number;
        startRegistrationDate?: string;
        endRegistrationDate?: string;
        maxStudent?: number;
        branchId?: number;
        skillCenterId?: number;
        courseSectionNumberTotal?: number;
        courseSectionNumber?: number;
        officialExamDate?: string;
        classPeriod?: number;
        examCourses?: any | null;
        certificateReviewBatchId?: number;
        program?: {
            skillCenterId?: number;
            programTypeId?: number;
            totalClassPeriodNumber?: number;
            totalHours?: number;
            isTesting?: boolean;
            certificateId?: number;
            isCancel?: boolean;
            note?: string;
            price?: number;
            scoreSystem?: number;
            scoreFormula?: number;
            scorePass?: number;
            testScoreSystem?: number;
            testScoreFormula?: number;
            testScorePass?: number;
            certificate?: any | null;
            skillCenter?: any | null;
            subjects?: any | null;
            programType?: any | null;
            programSubjects?: any[];
            scoreConfigs?: any | null;
            id?: number;
            code?: string;
            name?: string;
            concurrencyStamp?: string;
            isEnabled?: boolean;
            modifiedWhen?: string;
            modifiedBy?: number;
            modifiedFullName?: string;
        };
        branch?: any | null;
        roomType?: any | null;
        skillCenter?: any | null;
        certificateReviewBatch?: {
            certificateId?: number;
            conditionPass?: any | null;
            certificate?: {
                type?: number;
                link?: string;
                note?: string;
                skillCenterId?: number;
                skillCenter?: any | null;
                id?: number;
                code?: string;
                name?: string;
                concurrencyStamp?: string;
                isEnabled?: boolean;
                modifiedWhen?: string;
                modifiedBy?: number;
                modifiedFullName?: string;
            };
            id?: number;
            code?: string;
            name?: string;
            concurrencyStamp?: string;
            isEnabled?: boolean;
            modifiedWhen?: string;
            modifiedBy?: number;
            modifiedFullName?: string;
        };
        id?: number;
        code?: string;
        name?: string;
        concurrencyStamp?: string;
        isEnabled?: boolean;
        modifiedWhen?: string;
        modifiedBy?: number;
        modifiedFullName?: string;
    };
    certificateDecision?: any | null;
    id?: number;
    code?: string;
    name?: string;
    concurrencyStamp?: string;
    isEnabled?: boolean;
    modifiedWhen?: string;
    modifiedBy?: number;
    modifiedFullName?: string;
}